requirements.txt
pdfplumber>=0.7.6
python-dateutil>=2.8.2
typer>=0.9.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pytest>=7.0.0
streamlit